pyproject.toml
[project]
name = "lambda-converter"
version = "0.1.0"
description = ""
authors = [
    {name = "Troybado",email = "104241354+Troybado@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "pydantic (>=2.11.3,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "openai (>=1.76.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry.scripts]
lambda-convert = "lambda_converter.cli:app"
